python/requirements.txt
Flask==0.12.1
google-cloud-datastore==1.0.0
